core: compress timetables saved to s3
